requirements.txt
pytorch-lightning==1.0.3
pytorch-lightning-bolts==0.2.2
scipy==1.2.1
numpy==1.16.4
pandas==1.0.5
PyYAML==5.3.1
neptune-client==0.4.122
psutil==5.7.2
scikit-learn==0.23.2
h5py==2.10.0
matplotlib==3.3.2
seaborn==0.11.0